+++ CHANGELOG.md
@@ -2,10 +2,13 @@
 
 ## master
 * Add email settings link
+* Begin selecting payments for processing
 * Enable activation and version checking
+* Remove donate links
 * Removed unused EDD_Retroactive_Licensing_Settings
 * Update description
 * Use `get_edd_options` and `set_edd_options` methods
+* Verbiage updates
 
 ## 0.0.1
 * Initial code release